fix: vehicle update now deletes images not selected

+++ src/main/java/br/ufscar/dc/dsw/controller/VehicleController.java
@@ -108,9 +108,12 @@ public class VehicleController {
                 }
             }
 
-            for (String imageUrl : vehicle.getImages()) {
-                if (!finalImages.contains(imageUrl)) {
-                    storageService.delete(imageUrl);
+            Vehicle vehicleToUpdate = vehicleService.findById(vehicle.getId());
+            if (vehicleToUpdate != null && vehicleToUpdate.getImages() != null) {
+                for (String imageUrl : vehicleToUpdate.getImages()) {
+                    if (!finalImages.contains(imageUrl)) {
+                        storageService.delete(imageUrl);
+                    }
                 }
             }
